Move each command definitions to separate modules
Just to keep things clean.
