+++ dsync-cli/src/cli/command/model/host.rs
@@ -1,0 +1,11 @@
+use clap::Subcommand;
+
+#[derive(Subcommand)]
+pub(crate) enum HostCommand {
+    List {
+        /// Run host discovery before returning the result
+        #[arg(short = 'd', long = "discover")]
+        discover: Option<bool>,
+    },
+    Discover,
+}
